+++ CHANGELOG.md
@@ -5,7 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.5] - 2025-1-21
+## [1.1.6] - 2025-09-13
+
+### Added
+
+- Include Diagrammer.Core module as a requirement.
+- Introduce the necessary file structure to enable diagram generation.
+
+### Changed
+
+- Update AsBuiltReport.Core module requirement to v1.4.3.
+- Upgrade actions/checkout to v5 in Codeql.yml and PSScriptAnalyzer.yml.
+- Upgrade actions/bluesky-post-action to v0.3.0
+- Modifie Release.yml to use windows-latest.
+- Enhance Invoke-ASBuiltReport function with improved dependency version checks.
+- Update CHANGELOG.md to reflect recent changes.
+- This is the final release supporting VMware Horizon; future versions will support Omnissa only.
+
+## [1.1.5.1] - 2025-03-13
+
+### Fixed
+
+- Fix `Get-RequiredModule` script function to properly check for installed VMware PowerCLI versions ([Fix #36](https://github.com/AsBuiltReport/AsBuiltReport.VMware.Horizon/issues/36))
+
+## [1.1.5] - 2025-01-21
 
 ### Added